Native Python Distutils: Prevent searching host libraries
This change stops Distutils from adding the LIBDIR value from Python's
installed config/Makefile to its library_dirs list which is used to build the
linker command line. Presumably this was intended to be the directory to find
the Python library, but in my usage of Open embedded that library is
automatically found in the sysroot directory without any explicit specifying
a search directory.

Previously building pycairo-1.8.0 was failing with the following:

arm-angstrom-linux-gnueabi-gcc . . . -L/usr/lib -lcairo -lpython2.6 . . .
/usr/lib/libgcc_s.so.1: file not recognized: File format not recognized