Fix multiline expression comments causing errors in Bit Bake 1.14
Example:
ERROR: There is a comment on line 10 of file /media/disk/home/proj/aegle/openembedded/recipes/images/e-image.bb (#                        task-gpe-apps \) which is in the middle of a multiline expression.
Bitbake used to ignore these but no longer does so, please fix your metadata as errors are likely as a result of this change.